Refactor custom proxy form out of Api access method component
